Implement enable/disable aix-soname options.
Support filename-based shared library versioning on AIX with the lib.so
library filename extension, which is used with runtime linking only.
Runtime linking is enabled by the -brtl linker flag for executables and
the -G linker flag for Shared Objects.
(aix-soname=no): Default, also known as --disable-aix-soname.
  This is the current situation.
  (lib.a): When the -brtl linker flag is found in LDFLAGS, create from
    static objects only. Otherways, create as traditional shared library
    archive from the Shared Object built without runtime linking, with
    in-archive versioning.
  (lib.so): When the -brtl linker flag is found in LDFLAGS, create as
    the standalone Shared Object built with the -G linker flag, without
    any versioning support. Otherways, do not create it at all.
  (executables): When the -brtl linker flag is found in LDFLAGS, prefer
    to link against libraries with the .so filename extension.
    Otherways, link against libraries with the .a filename extension.
  (module): When the -brtl linker flag is found in LDFLAGS, same as
    lib.so.  Otherways, same as lib.a.
(aix-soname=yes): This is the new situation.
  (lib.a): Create from static objects.
  (lib.so): Create as shared library archive from the Shared Object
    named shr.o or shr_64.o, built with the -G linker flag, and the
    F_LOADONLY flag set, plus an Import File named shr.imp or
    shr_64.imp, respectively, necessary to allow for filename-based
    versioning.
  (executables): Add the -brtl linker flag to LDFLAGS, to prefer linking
    against shared libraries with the .so filename extension.
  (module): Same as lib.so.
* m4/libtool.m4: Support the "aix-soname".
* libltdl/ltdl.c: (LT_SHARED_LIB_MEMBER) When not defined by configure
because of --disable-aix-soname, still define for AIX again to allow for
loading foreign libtool archives specifying an archive member.
* libltdl/loaders/dlopen.c: (vm_open) When LT_SHARED_LIB_MEMBER is
defined, and we are advised to try known extensions (from lt_dlopenext),
and no shared archive member was specified, and dlopen fails with "Exec
format error", indicating the requested file was found but cannot be
loaded as standalone Shared Object, try again with specifying the
default shared archive member.